+++ src/main/java/br/com/luizfelipeduarte/imobiliariaapi/repositories/CidadeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import br.com.luizfelipeduarte.imobiliariaapi.entidade.Cidade;
 
 @Repository
-public interface CidadeRepository extends JpaRepository<Cidade,Integer>{
+public interface CidadeRepository extends JpaRepository<Cidade,Long>{
 	
 	public Cidade findByNome(String nome);
