+++ src/main/java/br/com/luizfelipeduarte/imobiliariaapi/repositories/EstadoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import br.com.luizfelipeduarte.imobiliariaapi.entidade.Estado;
 
 @Repository
-public interface EstadoRepository extends JpaRepository<Estado,Integer>{
+public interface EstadoRepository extends JpaRepository<Estado,Long>{
 	
 	public Estado findByNome(String nome);
